fingerprint: fpc: boost cpufreq on unlock
To reduce wake-up latency when using the fingerprint scanner we
register an input device and send a "key press" when the scanner
is being used during screen off. This will boost the cpus through
cpu-boost and reduce the screen on latency.

On touch the fp sensor, boost the cpu even if screen is on - psndna88@xda